refactor: package modules under scriptoria and add CI
