+++ CONTRIBUTING.md
@@ -1,0 +1,19 @@
+# Contributing to Scriptoria
+
+Thank you for contributing!
+
+## Run tests
+
+Install dependencies and the package in editable mode:
+
+```bash
+pip install -r requirements.txt
+pip install -e .
+```
+
+Then run the test suite:
+
+```bash
+pytest -q
+```
+
