+++ scriptoria/__init__.py
@@ -1,0 +1,1 @@
+"""Scriptoria package providing file management utilities."""
