+++ setup.py
@@ -1,0 +1,2 @@
+from setuptools import setup, find_packages
+setup(name='scriptoria', version='0.1.0', packages=find_packages())
